Refactor network tests
